project1/requirements.txt
gradio
openai
torch
transformers
torchaudio
python-dotenv

